Cli option -P now preserves the gamma ramps
Before version 1.12 you could set the temperature color
with -O cli option with with out any extra arguments.
With 1.12 -P for reseting the gamma ramps and it is
confusing because before, you didn't need that.

For that reason this patch makes the -P option to
preserver the gamma ramps inorder to make the cli
interface more consistent across the versions